[nrf fromlist] modules: mbedtls: Replace select statement with depends on for EC
Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/51001

Instead of using "select" on certain EC configurations, which is
considered unsafe for various reasons, use a "depends on" and rely on
the user to set a proper configuration in the config file.

Update the respective project configurations to comply with the new
configuration scheme.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>